Maximum Float Value

+++ src/com/bridgelabz/MaxValue.java
@@ -2,7 +2,7 @@ package com.bridgelabz;
 
 public class MaxValue {
 
-    void maxValue(Integer num1, Integer num2, Integer num3) {
+    void maxValue(Float num1, Float num2, Float num3) {
 
         if ((num1.compareTo(num2)) >= 0 && (num1.compareTo(num3)) >= 0)
             System.out.println(num1 + " is the largest number");
